refactor: custom pb types to own module
